- Add multi-valued SBinary_short support to mapi_SPropValue_CTR in the IDL
- Add GetDefaultFolder implementation to simple_mapi.c. This function
provides a convenient way to retrieve default folders ID. const
olFolder values are stored in libmapi/mapidefs.h

- Modify the torture test suite to reflect these changes.

- OpenMsgStore now decodes all the fid returned in the response

- fix a bug size in libmapi/property.c

+++ libmapi/property.c
@@ -313,7 +313,7 @@ _PUBLIC_ uint32_t cast_mapi_SPropValue(struct mapi_SPropValue *mapi_sprop, struc
 	switch(sprop->ulPropTag & 0xFFFF) {
 	case PT_BOOLEAN:
 		mapi_sprop->value.b = sprop->value.b;
-		return sizeof(uint16_t);
+		return sizeof(uint8_t);
 	case PT_I2:
 		mapi_sprop->value.i = sprop->value.i;
 		return sizeof(uint16_t);
